requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
PyMySQL==1.1.0
mysql-connector-python==8.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
ollama>=0.1.6
faiss-cpu>=1.12.0
sentence-transformers==5.1.1
langgraph==0.6.8